Addressed PR comments: removed comments, moved sac models to separate file
